Assistant: move installPythonPackage tool to python extension
Also tag it as requiring actions so it is not available in Ask mode, and tag it with requires-session:python so it only shows up when a Python session is active (though it actually uses the Terminal to install the packages, not the Console)
