+++ extensions/positron-python/package.json
@@ -1911,6 +1911,33 @@
         ],
         "languageModelTools": [
             {
+                "name": "installPythonPackage",
+                "displayName": "Install Python Package",
+                "modelDescription": "Install Python packages using pip. Provide an array of package names to install.",
+                "icon": "$(package)",
+                "toolReferenceName": "installPythonPackage",
+                "userDescription": "%python.languageModelTools.installPythonPackage.userDescription%",
+                "canBeReferencedInPrompt": true,
+                "tags": [
+                    "positron-assistant",
+                    "requires-session:python",
+                    "requires-actions"
+                ],
+                "inputSchema": {
+                    "type": "object",
+                    "properties": {
+                        "packages": {
+                            "type": "array",
+                            "description": "Array of Python package names to install.",
+                            "items": {
+                                "type": "string",
+                                "description": "Name of the Python package to install."
+                            }
+                        }
+                    }
+                }
+            },
+            {
                 "name": "get_python_environment_details",
                 "displayName": "Get Python Environment Info",
                 "when": "false",
