+++ extensions/positron-python/package.nls.json
@@ -2,6 +2,7 @@
     "displayName": "Python",
     "description": "Python kernel, Debugging (multi-threaded, remote), code formatting, refactoring, unit tests, and more.",
     "python.command.python.startTerminalREPL.title": "Start Terminal REPL",
+    "python.languageModelTools.installPythonPackage.userDescription": "Install Python packages in the current environment",
     "python.languageModelTools.get_python_environment_details.userDescription": "Get information for a Python Environment, such as Type, Version, Packages, and more.",
     "python.languageModelTools.install_python_packages.userDescription": "Installs Python packages in a Python Environment.",
     "python.languageModelTools.get_python_executable_details.userDescription": "Get executable info for a Python Environment",
